Fix race condition when asking for instances
The worker server would add instances to the state based on the
connection_ready message.  At this point, the connection is ready but the
worker_instance information is not available.  The change introduces a
new message which is sent after the worker_instance information has been
collected.  The worker server now uses this message to add instances to
the list of instances it manages.